+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unreleased
+- [FEATURE] Add `currentSessionID(completion:)` accessor to access the current session ID.
 - [FEATURE] Add `BatchProcessingLevel` configuration allowing to process more batches within single read/upload cycle. See [#1531][]
 - [FIX] Use `currentRequest` instead `originalRequest` for URLSession request interception